Add files via upload
It’s amazing how much stuff that comes to your attention once you start focusing on something - anything - all upcoming details... Initially the intention was only to test some features but ended up as an all only javaScript page! Everything is built and rendered with JS except for some necessary HTML. It was really a good way to explore and test. Yes, the traditional way of split a site into built, styling and functionality is to prefer and follow; but I sort of got into this and find it easier in some way only following one method all the way through. One thing is for sure true: JavaScript is Fun!